refactor(sdk): extract telemetry runtime; keep attributes internal, introduce runtime-accessors
Phase B.2 of ses_004-sdk-dependency-diet. SDK-side companion to the
B.1 move into @namzu/telemetry.

Deleted:
- packages/sdk/src/provider/telemetry/setup.ts (moved to telemetry)
- packages/sdk/src/telemetry/index.ts (unused barrel after runtime moves)
- packages/sdk/src/telemetry/metrics.ts (moved to telemetry)
- packages/sdk/src/types/telemetry/ (types moved to telemetry)

Added:
- packages/sdk/src/telemetry/runtime-accessors.ts — ~18-line
  SDK-internal getTracer()/getMeter() readers that resolve against
  @opentelemetry/api globals. Not re-exported by the root barrel.

Rewired:
- 3 span-starting call sites (runtime/query/index.ts,
  runtime/query/iteration/index.ts, registry/tool/execute.ts) now import
  getTracer from telemetry/runtime-accessors.js instead of the deleted
  provider/telemetry/setup.js. Import order auto-fixed by biome.

Root barrel (src/index.ts):
- Removed 10-line telemetry runtime export block:
  TelemetryProvider, initTelemetry, getTelemetry, getTracer, getMeter,
  createPlatformMetrics, PlatformMetrics, and `export *
  from './telemetry/attributes.js'`.
- Removed `export * from './types/telemetry/index.js'`. TelemetryConfig
  and ExporterType now ship from @namzu/telemetry only.

Retained:
- packages/sdk/src/telemetry/attributes.ts — SDK-internal. The 4 span
  call sites import GENAI/NAMZU/span-name helpers from here. No longer
  exposed by the public barrel; external consumers use
  @namzu/telemetry/attributes.

SDK package.json still declares the 10 runtime deps; Phase B.3 moves
them to peer + devDep mirror. pnpm -r build + test green.
