+++ packages/sdk/src/runtime/query/iteration/index.ts
@@ -5,9 +5,9 @@ import type { WorkingStateManager } from '../../../compaction/manager.js'
 import type { CompactionConfig } from '../../../config/runtime.js'
 import type { PlanManager } from '../../../manager/plan/lifecycle.js'
 import type { RunPersistence } from '../../../manager/run/persistence.js'
-import { getTracer } from '../../../provider/telemetry/setup.js'
 import type { ActivityStore } from '../../../store/activity/memory.js'
 import { GENAI, NAMZU, agentIterationSpanName } from '../../../telemetry/attributes.js'
+import { getTracer } from '../../../telemetry/runtime-accessors.js'
 import type { ResumeHandler } from '../../../types/hitl/index.js'
 import { createAssistantMessage, createUserMessage } from '../../../types/message/index.js'
 import type { LLMProvider } from '../../../types/provider/index.js'
